LPS-63481 Move contexts to document-library-api
