+++ modules/apps/collaboration/document-library/document-library-api/src/main/java/com/liferay/document/library/display/context/DLFilePicker.java
@@ -12,12 +12,25 @@
  * details.
  */
 
-package com.liferay.document.library.kernel.display.context;
+package com.liferay.document.library.display.context;
 
-import com.liferay.portal.kernel.display.context.DisplayContext;
+import com.liferay.portal.kernel.exception.PortalException;
 
 /**
  * @author Iván Zaera
  */
-public interface DLDisplayContext extends DisplayContext {
+public interface DLFilePicker {
+
+	public String getDescriptionFieldName();
+
+	public String getIconFieldName();
+
+	public String getJavaScript() throws PortalException;
+
+	public String getJavaScriptModuleName();
+
+	public String getOnClickCallback();
+
+	public String getTitleFieldName();
+
 }
